Restore paper/legacy review switch pattern to fix E2E compatibility
